backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
audioop-lts==0.2.1
numpy==2.2.1
azure-identity==1.19.0
aiohttp==3.13.3
